split up singleton specs into their own testfiles
